requirements.txt
botcity-framework-web
botcity-framework-core
PyYAML